+++ run_tests.py
@@ -14,8 +14,8 @@ TEST_DIR = 'tests'
 if __name__ == "__main__":
     if coverage:
         cov = coverage.coverage(source=['github3'],
-                omit=['github3/packages/*'])
-        cov.exclude('(No coverage)')
+                                omit=['github3/packages/*'])
+        cov.exclude('\(No coverage\)')
         cov.exclude('def __repr__')
         cov.start()